New response button ("-"), clarify language, deleted face in tun_prac
Deleted one face in Tuning Practice to prevent drawing 4 faces in the practice run.

Added "-" as a possible response response to choose an image during retrieval (this is helpful if you don't have a number pad).

Adapted language (eng, ger, dutch) to clarify that an arrow key press is required to start the experiment, not just any key press.